`Fix`: Deploy crash when location coordinates are `NaN`.
The previous script was checking agains `null` so some of the cases
such as NaN, which can be more then just null slipped through (In
this case an empty string).

New code uses `Number()` instead of `parseFloat()` which returns
always `NaN` in case of incorrect conversion and check against
it in subsequent checks.